Move API param handling to separate file
